pyproject.toml
[project]
name = "comicizer-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.0",
    "google-generativeai>=0.8.5",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "reportlab>=4.4.5",
    "uvicorn>=0.38.0",
]
